Move enclave to own module
