+++ enclave/enclave-api/src/main/java/com/quorum/tessera/enclave/RawTransaction.java
@@ -1,8 +1,9 @@
-package com.quorum.tessera.encryption;
+package com.quorum.tessera.enclave;
 
+import com.quorum.tessera.encryption.PublicKey;
 import com.quorum.tessera.nacl.Nonce;
-import java.util.Arrays;
 
+import java.util.Arrays;
 import java.util.Objects;
 
 public class RawTransaction {
