+++ tessera-core/src/test/java/com/quorum/tessera/transaction/TransactionManagerTest.java
@@ -1,13 +1,9 @@
 package com.quorum.tessera.transaction;
 
 import com.quorum.tessera.api.model.*;
-import com.quorum.tessera.enclave.Enclave;
-import com.quorum.tessera.enclave.EncodedPayload;
-import com.quorum.tessera.enclave.EncodedPayloadWithRecipients;
-import com.quorum.tessera.enclave.PayloadEncoder;
+import com.quorum.tessera.enclave.*;
 import com.quorum.tessera.enclave.model.MessageHash;
 import com.quorum.tessera.encryption.PublicKey;
-import com.quorum.tessera.encryption.RawTransaction;
 import com.quorum.tessera.nacl.NaclException;
 import com.quorum.tessera.transaction.exception.TransactionNotFoundException;
 import com.quorum.tessera.transaction.model.EncryptedRawTransaction;
